+++ Changelog.md
@@ -129,3 +129,5 @@
 1.2.9 fix limit in subqueries
 1.3.0 fix order by in subqueries
 1.3.1 fix order by in subqueries
+1.3.2 fix order by in subqueries
+1.3.3 fix group by in subqueries